Moved over KaTeX preprocessor
