+++ src/lib.rs
@@ -7,7 +7,6 @@ pub mod log;
 
 pub mod cmd;
 pub mod config;
-pub mod katex;
 pub mod preprocess;
 pub mod renderer;
 pub mod utils;
