+++ src/preprocess/katex.rs
@@ -1,0 +1,153 @@
+use mdbook::book::Chapter;
+use pulldown_cmark::{Event, Options, Parser, Tag};
+
+const KATEX_CSS: &str = r#"<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/katex@0.13.18/dist/katex.min.css" integrity="sha384-zTROYFVGOfTw7JV7KUu8udsvW2fx4lWOsCEDqhBreBwlHI4ioVRtmIvEThzJHGET" crossorigin="anonymous">"#;
+const KATEX_JS: &str = r#"<script defer src="https://cdn.jsdelivr.net/npm/katex@0.13.20/dist/katex.min.js" integrity="sha384-ov99pRO2tAc0JuxTVzf63RHHeQTJ0CIawbDZFiFTzB07aqFZwEu2pz4uzqL+5OPG" crossorigin="anonymous"></script>"#;
+const RENDER_SCRIPT: &str = r#"<script>
+document.addEventListener('DOMContentLoaded', function() {
+    var math = document.querySelectorAll("[class^=katex]");
+    for (var i = 0; i < math.length; i++) {
+      // Convert escaped ampersands before rendering
+      var toRender = math[i].textContent.replace(/&amp;/g, '&');
+      katex.render(toRender, math[i]);
+    }
+}, false)
+</script>"#;
+
+pub fn render(ch: &mut Chapter) {
+    let copy = ch.content.clone();
+
+    let mut opts = Options::empty();
+    opts.insert(Options::ENABLE_TABLES);
+    opts.insert(Options::ENABLE_FOOTNOTES);
+    let parser = Parser::new_ext(&copy, opts);
+    let mut buf = String::new();
+    let mut render_math = false;
+    for event in parser {
+        match event {
+            // Respect formatting
+            Event::Start(Tag::Emphasis) | Event::End(Tag::Emphasis) => {
+                if render_math {
+                    buf.push('*')
+                }
+            }
+            Event::Start(Tag::Strong) | Event::End(Tag::Strong) => {
+                if render_math {
+                    buf.push_str("**")
+                }
+            }
+
+            // Clear buffer when code is hit
+            Event::Code(_) => buf.clear(),
+
+            // Don't render inside HTML
+            Event::Html(_) => render_math = !render_math,
+
+            // Start math rendering
+            Event::Start(Tag::Paragraph)
+            | Event::Start(Tag::Heading(_))
+            | Event::Start(Tag::List(_))
+            | Event::Start(Tag::Item)
+            | Event::Start(Tag::Table(_))
+            | Event::Start(Tag::TableHead)
+            | Event::Start(Tag::TableCell)
+            | Event::Start(Tag::BlockQuote) => {
+                render_math = true;
+            }
+            // If rendering math, push text events to buffer
+            Event::Text(text) => {
+                if render_math {
+                    buf.push_str(&text);
+                    if let Some(new_buf) = inline(&buf) {
+                        // If inline math is found within this text event, render it
+                        // and clear buffer.
+                        replace(ch, &buf, &new_buf);
+                        buf.clear()
+                    }
+                }
+            }
+            // Stop math rendering
+            Event::End(Tag::Paragraph)
+            | Event::End(Tag::Heading(_))
+            | Event::End(Tag::List(_))
+            | Event::End(Tag::Item)
+            | Event::End(Tag::Table(_))
+            | Event::End(Tag::TableHead)
+            | Event::End(Tag::TableCell)
+            | Event::End(Tag::BlockQuote) => {
+                if let Some(display) = display(&buf) {
+                    replace(ch, &buf, &display);
+                } else if let Some(new_text) = inline(&buf) {
+                    replace(ch, &buf, &new_text);
+                }
+                buf.clear();
+                render_math = false;
+            }
+
+            _ => {}
+        }
+    }
+    ch.content.push_str(&format!("\n\n{}", KATEX_CSS));
+    ch.content.push_str(&format!("\n\n{}", KATEX_JS));
+    ch.content.push_str(&format!("\n\n{}", RENDER_SCRIPT));
+}
+
+fn inline(text: &str) -> Option<String> {
+    let mut out = text.to_string();
+    let mut splits = text.split('$');
+    let mut escaped = match splits.next() {
+        Some(first_split) => first_split.ends_with('\\'),
+        None => return None, // No dollars, do early return
+    };
+
+    for split in splits {
+        if split.is_empty() {
+            escaped = true; // Two $s after each other. Do not consider inline.
+            continue;
+        } else if escaped
+            || split.starts_with(' ')
+            || split.ends_with(' ')
+            || split.chars().next().unwrap().is_numeric()
+        {
+            if split.ends_with('\\') {
+                escaped = true;
+            } else {
+                escaped = false;
+            }
+        } else if !text.ends_with(split) {
+            // Hacky way of checking if this is the last split
+            out = out.replacen(
+                &format!("${}$", split),
+                &format!("<span class=\"katex-inline\">{}</span>", split),
+                1,
+            );
+        }
+    }
+
+    if out != text {
+        Some(out)
+    } else {
+        None
+    }
+}
+
+fn display(text: &str) -> Option<String> {
+    if let Some(math) = text.strip_prefix("$$") {
+        if let Some(math) = math.strip_suffix("$$") {
+            return Some(text.replacen(
+                text,
+                &format!("<div class=\"katex-display\">{}</div>", math),
+                1,
+            ));
+        }
+    }
+    None
+}
+
+fn fix(text: impl std::ops::Deref<Target = str>) -> String {
+    text.replace(r#" \ "#, r#" \\ "#)
+}
+
+fn replace(ch: &mut Chapter, old: &str, new: &str) {
+    ch.content = ch.content.replacen(&fix(old), &fix(new), 1);
+}
